+++ components/atoms/Home/Products.tsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect } from "react";
 import { View, Text, StyleSheet, Image } from "react-native";
 import { FlatList, TouchableOpacity } from "react-native-gesture-handler";
 import { useNavigation } from "@react-navigation/native";
 import { useSelector, useDispatch } from "react-redux";
 import { fetchProducts } from "../../../reducks/products/operations";
 
-export default function Products(props) {
+export default function Products() {
   const { navigate } = useNavigation();
 
   const selector = useSelector((state) => state);
@@ -14,28 +14,12 @@ export default function Products(props) {
   const products = selector.products;
   const list = products.list;
 
-  // const value = props.value;
-  // const [items, setItems] = useState();
-
   useEffect(() => {
     dispatch(fetchProducts());
   }, []);
 
-  // useEffect(() => {
-  //   setItems(list);
-  // }, [list]);
-
-  // useEffect(() => {
-  //   const updateList = list.filter((item) => {
-  //     return item.name.toLowerCase().indexOf(value.toLowerCase()) !== -1;
-  //   });
-
-  //   setItems(updateList);
-  // }, [value]);
-
   return (
     <FlatList
-      // data={items}
       data={list}
       contentContainerStyle={styles.scrollView}
       keyExtractor={(item) => `products-${item.id}`}
