+++ components/atoms/Home/index.tsx
@@ -5,4 +5,5 @@ export { default as Product } from "./Product";
 export { default as Products } from "./Products";
 export { default as Purchase } from "./Purchase";
 export { default as Search } from "./Search";
+export { default as SearchHome } from "./SearchHome";
 export { default as SearchScreen } from "./SearchScreen";
