+++ components/pages/Home.tsx
@@ -4,17 +4,14 @@ import { createStackNavigator } from "@react-navigation/stack";
 import { createMaterialTopTabNavigator } from "@react-navigation/material-top-tabs";
 import { useNavigation } from "@react-navigation/native";
 import Icon from "react-native-vector-icons/FontAwesome";
-import { Category, Products, Search, SearchScreen } from "../atoms/Home";
+import { Category, Products, Search, SearchHome } from "../atoms/Home";
 
 const Tab = createMaterialTopTabNavigator();
 const Stack = createStackNavigator();
 
-function SearchTab({ value }) {
+function SearchTab() {
   return (
     <Tab.Navigator tabBarOptions={{ scrollEnabled: true }}>
-      {/* <Tab.Screen name="おすすめ">
-        {(props) => <Products {...props} value={value} />}
-      </Tab.Screen> */}
       <Tab.Screen name="おすすめ" component={Products} />
       <Tab.Screen name="新着" component={Products} />
       <Tab.Screen name="カテゴリー" component={Category} />
@@ -31,35 +28,14 @@ export default function Home() {
     <Stack.Navigator>
       <Stack.Screen
         name="Home"
+        component={SearchTab}
         options={{
           headerTitle: () => (
-            // <TextInput
-            //   style={styles.input}
-            //   onChangeText={(newValue) => {
-            //     setValue(newValue);
-            //   }}
-            //   placeholder="検索"
-            //   autoCapitalize="none"
-            // />
             <TouchableOpacity
               style={styles.input}
-              // onChangeText={(newValue) => {
-              //   setValue(newValue);
-              // }}
-              // placeholder="検索"
-              // autoCapitalize="none"
               onPress={() => navigate("Search")}
             >
-              <Text
-                style={{
-                  color: "#555",
-                  justifyContent: "center",
-                  alignItems: "center",
-                  marginTop: 8,
-                }}
-              >
-                なにをお探しですか?
-              </Text>
+              <Text style={styles.searchText}>なにをお探しですか?</Text>
             </TouchableOpacity>
           ),
           headerLeft: () => <Icon name="qrcode" size={24} />,
@@ -67,10 +43,9 @@ export default function Home() {
           headerLeftContainerStyle: styles.headerLeft,
           headerRightContainerStyle: styles.headerRight,
         }}
-      >
-        {(props) => <SearchTab value={value} />}
-      </Stack.Screen>
-      {/* <Stack.Screen
+      />
+
+      <Stack.Screen
         name="Search"
         component={Search}
         options={{
@@ -82,52 +57,15 @@ export default function Home() {
               }}
               placeholder="検索"
               autoCapitalize="none"
-              onSubmitEditing={()=>navigate('SearchScreen')}
+              onSubmitEditing={() => navigate("SearchHome")}
             />
           ),
           headerBackTitleVisible: false,
         }}
-      /> */}
+      />
+
       <Stack.Screen
-        name="Search"
-        // component={Search}
-        options={{
-          headerTitle: () => (
-            <TextInput
-              style={styles.input}
-              onChangeText={(newValue) => {
-                setValue(newValue);
-              }}
-              placeholder="検索"
-              autoCapitalize="none"
-              onSubmitEditing={()=>navigate('SearchScreen')}
-            />
-          ),
-          headerBackTitleVisible: false,
-        }}
-      >
-        {(props) => <Search value={value} />}
-      </Stack.Screen>
-      {/* <Stack.Screen
-        name="SearchScreen"
-        component={SearchScreen}
-        options={{
-          headerTitle: () => (
-            <TextInput
-              style={styles.input}
-              onChangeText={(newValue) => {
-                setValue(newValue);
-              }}
-              placeholder="検索"
-              autoCapitalize="none"
-            />
-          ),
-          headerBackTitleVisible: false,
-        }}
-      /> */}
-      <Stack.Screen
-        name="SearchScreen"
-        // component={SearchScreen}
+        name="SearchHome"
         options={{
           headerTitle: () => (
             <TextInput
@@ -143,7 +81,7 @@ export default function Home() {
           headerBackTitleVisible: false,
         }}
       >
-        {(props) => <SearchScreen value={value} />}
+        {() => <SearchHome value={value} />}
       </Stack.Screen>
     </Stack.Navigator>
   );
@@ -165,5 +103,11 @@ const styles = StyleSheet.create({
     borderRadius: 20,
     marginRight: 20,
     marginLeft: 20,
+  },
+  searchText: {
+    color: "#555",
+    justifyContent: "center",
+    alignItems: "center",
+    marginTop: 8,
   },
 });
